Validation of users
In this, you can validate the user from the 'https://reqres.in/api/users' link and provide the username as a first_name and the password should be last_name. Click to login button can validate you. And if validated this goes you to the profile.html and greetings the name and click to Log Out goes to the login page again.